IPv6 support for rblsmtpd
Needs a tcpserver with IPv6 support that set TCP6 env variable.

This is a patch against rblsmtpd.c that bypasses the DNSBL checks if the connection is IPv6 based.

This is simply due to the fact that there are currently no IPv6 enabled DNSBLs and looking up an IPv6 address has not been specified.

In the future the patch could be changed to reverse the IPv6 address in the nibble format used for ip6.arpa. lookups.