http://gurubert.de/qmail/rblsmtpd/message-ip
This is a patch against rblsmtpd from ucspi-tcp 0.88 that allows to set a $RBLMESSAGE environment variable (e.g. in /etc/tcp.smtp).

The content of the variable will be prepended to the string that is returned by querying the various DNSBLs.

This can be used e.g. to add contact info to the SMTP error message the sender receives.

The patch also replaces %IP% in the DNSBL return string with the value of $TCPREMOTEIP. Some DNSBLs do return an URL with %IP% as a placeholder.